Fix Binance Client tests
