+++ api/deals/controllers.py
@@ -19,8 +19,10 @@ from tools.handle_error import NotEnoughFunds, QuantityTooLow, handle_binance_er
 from tools.round_numbers import round_numbers, supress_notation
 from scipy.stats import linregress
 
-
 class CreateDealControllerError(Exception):
+    pass
+
+class StreamingSaveError(Exception):
     pass
 
 class CreateDealController(BaseDeal):
@@ -61,8 +63,36 @@ class CreateDealController(BaseDeal):
             return None
         qty = round_numbers(balance, self.qty_precision)
         return qty
-    
 
+    def save_bot_streaming(self):
+        """
+        MongoDB query to save bot using Pydantic
+
+        This function differs from usual save query in that
+        it returns the saved bot, thus called streaming, it's
+        specifically for streaming saves
+        """
+
+        try:
+
+            bot = encode_json(self.active_bot)
+            if "_id" in bot:
+                bot.pop("_id")
+
+            bot = self.db_collection.find_one_and_update(
+                {"id": self.active_bot.id},
+                {
+                    "$set": bot,
+                },
+                return_document=ReturnDocument.AFTER,
+            )
+
+        except Exception as error:
+            self.update_deal_logs(f"Failed to save bot during streaming updates: {error}")
+            raise StreamingSaveError(error)
+
+        return bot
+    
     def base_order(self):
         """
         Required initial order to trigger bot.
@@ -257,7 +287,12 @@ class CreateDealController(BaseDeal):
             if not qty:
                 closed_orders = self.close_open_orders(self.active_bot.pair)
                 if not closed_orders:
-                    print(f"Bot already closed? There is no {self.active_bot.pair} quantity in the balance. Please delete the bot.")
+                    self.update_deal_logs(
+                        f"No quantity in balance, no closed orders. Cannot execute update trailling profit."
+                    )
+                    self.active_bot.status = "error"
+                    self.save_bot_streaming()
+                    return
 
         # Dispatch fake order
         if self.db_collection.name == "paper_trading":
@@ -312,24 +347,7 @@ class CreateDealController(BaseDeal):
         self.active_bot.errors.append(msg)
         print(msg)
 
-        try:
-
-            bot = encode_json(self.active_bot)
-            if "_id" in bot:
-                bot.pop("_id")
-
-            bot = self.db_collection.find_one_and_update(
-                {"id": self.active_bot.id},
-                {
-                    "$set": bot,
-                },
-                return_document=ReturnDocument.AFTER,
-            )
-
-        except Exception as error:
-            self.update_deal_logs(f"Failed to close trailling take profit: {error}")
-            raise TraillingProfitError(error)
-
+        bot = self.save_bot_streaming()
         return bot
 
     def close_all(self):
@@ -627,8 +645,10 @@ class CreateDealController(BaseDeal):
             closed_orders = self.close_open_orders(self.active_bot.pair)
             if not closed_orders:
                 self.update_deal_logs(
-                    f"Cannot execute update stop limit, quantity is {qty}"
+                    f"No quantity in balance, no closed orders. Cannot execute update stop limit."
                 )
+                self.active_bot.status = "error"
+                self.save_bot_streaming()
                 return
 
         order_id = None
@@ -699,28 +719,8 @@ class CreateDealController(BaseDeal):
         self.active_bot.errors.append(msg)
         self.active_bot.status = "completed"
 
-        try:
-
-            bot = encode_json(self.active_bot)
-            if "_id" in bot:
-                bot.pop("_id")
-
-            self.db_collection.update_one(
-                {"id": self.active_bot.id},
-                {"$set": bot},
-            )
-
-        except ValidationError as error:
-            self.update_deal_logs(f"Stop loss error: {error}")
-            return
-        except (TypeError, AttributeError) as error:
-            message = str(";".join(error.args))
-            self.update_deal_logs(f"Stop loss error: {message}")
-            return
-        except Exception as error:
-            self.update_deal_logs(f"Stop loss error: {error}")
-            return
-        pass
+        bot = self.save_bot_streaming()
+        return bot
 
     def execute_short_sell(self):
         """
@@ -955,29 +955,7 @@ class CreateDealController(BaseDeal):
                     self.active_bot.deal.take_profit_price = float(close_price) + (float(close_price) * volatility)
                     self.active_bot.deal.trailling_profit_price = float(close_price) + (float(close_price) * volatility)
 
-            try:
-
-                bot = encode_json(self.active_bot)
-                if "_id" in bot:
-                    bot.pop("_id")
-
-                self.db_collection.update_one(
-                    {"id": self.active_bot.id},
-                    {"$set": bot},
-                )
-
-            except ValidationError as error:
-                self.update_deal_logs(f"Dynamic take profit error: {error}")
-                return
-            except (TypeError, AttributeError) as error:
-                message = str(";".join(error.args))
-                self.update_deal_logs(f"Dynamic take profit error: {message}")
-                return
-            except Exception as error:
-                self.update_deal_logs(f"Dynamic take profit error: {error}")
-                return
-
-        bot = encode_json(self.active_bot)
+        bot = self.save_bot_streaming()
         return bot
 
     def open_deal(self):
